backend/requirements.txt
psycopg2-binary
django-cors-headers
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
yfinance
